requirements.txt
Django==4.2.2
djangorestframework==3.14.0
pinecone-client==2.2.2
python-dotenv==1.0.0
django-cors-headers==4.1.0
dj-database-url==2.0.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
